Create absolute references for flags
When resolving fixture templates, what is required to find fixtures in the bag are fixture IDs and not (fixture) references. However the flag parser has no idea of which fixture flags it is parsing. As the parser is re-used for parsing FQCN for example, case where there is no fixture, I'm not sure it's a good idea to make it "fixture aware". An alternative solution is to reworking `Templating`, i.e. when creating a template fixture, to change the extended fixture references at that time. The only drawbacks I see is that it makes `FixtureReference` a bit confusing as it can refer to a relative OR absolute fixture, but in both cases they are accessed via `::getReference()` which comes from the implemented interface.