Change SpecificationBag merge behaviour (#400)
As of now, when merging a fixture specs with another, the overlapping values where overridden. However, when fixtures are being resolved, we start from the child fixtures and not the parent one. As a result, it is much more convenient to have the opposite behaviour which is keeping the existing values. Reversing the order of the resolution would be doable, but much more complex and less performant, whereas this fix is quite simple although a bit confusing.

Also fixed a bug with `MethodCallBag`: overlapping methods were being overridden instead of being stacked.