Add tests for templates declared in includes (#424)
As reported in https://github.com/nelmio/alice/issues/276, it appears that if a fixture extends a template declared in another file which is included, an exception stating that the template has not been found is thrown.

The issue has been reported the Sep 24, 2015 which is quite old at the time of writing. This commit includes a test to account for that scenario, which appears to have been solved since then.