[V3] Add expression language lexer (#434)
First part to tackle https://github.com/nelmio/alice/issue/#377: have a lexer to parse the inputs. This is the first step, the second one will be to have a Parser to transform the tokens in a more comprehensible set of data. The lexer + parser step should be added to the FixturesGenerator as everything is static, hence catchable.