[On Hold] Add support for stdClass (#425)
Sometimes, especially for testing, it is cumbersome to have to create a dedicated model and it would be nicer to have a more flexible class with the most basic behaviour. Ideally, stdClass would be a perfect candidate for that. However this cannot work as alice Populator will throw an exception as it will find no properties or methods. To avoid weird behaviours in regular usage, it is better to not change that and instead have a custom class to mimicate stdClass.

Closes https://github.com/nelmio/alice/issues/187