Remove immutability (#505)
The object graph generated by alice should not be immutable. As such immutability has been removed for the objects related to the objects described by the fixtures. Some elements like the Fixture properties or the method call are generated by alice and potentially requires mutability as well. As those elements are internal to alice (unless you dig really deep in which case you ought to know what you are doing), the risk presented by mutability are relatively small.