Better handling of references to inexistent fixtures
As of now if a reference to an inexistent fixture in a constructor is made, as no object is found for it, the reference resolver will try to instantiate that fixture. However as it is a temporary fixture, it has no class leading to this obscure error:

`Class  does not exist in /path/to/alice/src/Generator/Instantiator/Chainable/NullConstructorInstantiator.php:36`

A solution is to check if this fixture exist when no object is found to thrown an error before trying to generate it.