Automatically escape @ character in words
Closes #516.

I also took the opportunity of this change to slighyly change the wiring of the
lexer to favor the decorater pattern rather than the registry patterns. This make
it easier to add this new lexer, make the system easier to understand and remove
the usage of a compiler pass in the Symfony bridge.