Fix complete generation process
Fix of a major bug putted in light by the new tests in LoaderIntegrationTest: referencing to a property or a call on a fixture that is not fully generated yet.

Solving this problem slightly complexify the generation of the objects and brings the notion of 1. requiring to fully generate an object in one go and 2. introduce the notion of "complete object". This complexification is entirely justified in the sense that it is necessary to fulfil the promised made in #257 which is that the fixture order will have no importance in 3.x.