Fix unique scope for arrays (#634)
When we have the following:

```yaml
Dummy:
  dummy{1..10}:
    name (unique): <name()>
    relatedDummy (unique): 2x @dummy*
```

The  value for the `name` property will be unique for all the fixtures of the `Dummy` class. However, for `relatedDummy`, the behaviour of unique is quite different: we expect the scope of the uniqueness to be not the `Dummy` class, but each specific fixture `relatedDummy`, i.e. we just ensure that the array does not have any duplicates.