Add a context cache
Add a cache in the GenerationContext. The advantages of that approach is that the cache in question has the same lifespan as the context itself, i.e. a loading. This allow resolvers to remain stateless (unlike the current `FixtureWildcardReferenceResolver`) but still leverage caching. Having another cache pool could be useful but I fear too much side-effects.