Make testsuite compatible with PHPUnit 8 (#1023)
* Added void return type in PHPUnit hook methods

This is compatible with PHPUnit 7 (as return types can be added in a
child class), and prepares the support for PHPUnit 8+
Visibity of setup and tearDown is also updated to match the PHPUnit
visibility.

* Migrate away from expectedException annotations

This way of expecting exceptions is deprecated in PHPUnit 8.

* Remove broken test case in the provider

The test has a string typehint, so passing an integer value from the
data provider does not make any sense. The test was still testing with a
string. And when migrating to PHPUnit 8, things break due to PHPUnit 8
running in strict mode when calling the test.

* Fix invalid test data providers

PHPUnit 8 reports invalid data providers which define the same key
twice, instead of silently ignoring duplicates.
In some cases, these were indeed actual duplicates (probably from a bad
conflict resolution). In other cases, this was causing some cases to be
ignored.

* Allow using PHPUnit 8 to run tests